Add data migration to populate product categories
